requirements.txt
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
Pillow==10.0.0
python-dotenv==1.0.0
